Added --compressed, fixt parsing bug
added the --compressed flag, to generate csv-table, where similar components are tied together into one line

fixed a major bug on parsing the .net file. since we look only for "(lib " or "(ref ", the fields could get
overwritten, by this tags, which can occure outside of the group export:components:comp

TODO: the references in the compressed format are not sorted, which would look nice.